+++ src/mixture/cluster.jl
@@ -1,11 +1,11 @@
 mutable struct Cluster#{Theta}
-    number::Int64
-    logmass::Union{Float64, Void}
+    number::Int
+    logmass::Union{Float, Void}
     parameter::Hierarchy#Theta
-    w::Float64
+    w::Float
 end
 
-Cluster(number::Int64, logmass::Float64, parameter::Hierarchy) =
+Cluster(number::Int, logmass::Float, parameter::Hierarchy) =
 Cluster(number, logmass, parameter, 0.0)
 
 function isEmpty(cc::Cluster)
